Move PlayedCards to own component
